organize fn and file names of all the muster fns
