+++ qcdb/programs/cfour/harvester.py
@@ -1,12 +1,8 @@
-import uuid
 import struct
 
 import numpy as np
 import qcelemental as qcel
 from qcelemental.models import Molecule
-
-from ...exceptions import ValidationError
-from ...util import conv_float2negexp
 
 #   Ways of modifying a computation
 #   global:     set global c-side option
@@ -25,54 +21,6 @@ from ...util import conv_float2negexp
 #   kwarg ~= i-local > local > global > driver-def > i-def
 
 #   P4 infrastructure replacing interfaced infrastructure (mol, basis, mem) where unavoidable overlap in how things are specified (mult in mol{} vs keyword) is treated as a clobber & complain if conflict VS P4 infrastructure as an aliased/convenient leak into interfaced infrastructure (psi) and is strictly no clobber or complain.
-
-
-def muster_inherited_keywords(ropts: 'Keywords', verbose: int = 1) -> None:
-    """Translate psi4 keywords *opt* that have been explicitly set into
-    their Cfour counterparts. Since explicitly set Cfour module keyword
-    values will always be used preferentially to these inferred from
-    psi4, the 'clobber' property is set to False.
-
-    """
-    import sys
-    accession = sys._getframe().f_code.co_name + '_' + str(uuid.uuid4())[:8]
-    kwgs = {'accession': accession, 'verbose': verbose}
-    do_translate = ropts.scroll['QCDB']['TRANSLATE_QCDB'].value
-
-    # qcdb/memory [B] --> cfour/memory_size [MB]
-    qopt = ropts.scroll['QCDB']['MEMORY']
-    if do_translate or qopt.is_required():
-        mem = int(0.000001 * qopt.value)
-        print('\n\nMEMORY', mem, '\n\n')
-        ropts.suggest('CFOUR', 'MEMORY_SIZE', mem, **kwgs)
-        ropts.suggest('CFOUR', 'MEM_UNIT', 'MB', **kwgs)
-
-    # qcdb/puream --> cfour/spherical
-    ropts.suggest('CFOUR', 'SPHERICAL', ropts.scroll['QCDB']['PUREAM'].value, **kwgs)
-
-    # qcdb/reference --> cfour/reference
-    # TODO ref or scf__ref?
-    qref = ropts.scroll['QCDB']['SCF__REFERENCE'].value
-    if qref in ['RHF', 'UHF', 'ROHF']:
-        #ref = {'RHF': 'RHF',
-        #       'UHF': 'UHF',
-        #       'ROHF': 'ROHF'}[ropts.scroll['QCDB']['REFERENCE'].value]
-        ropts.suggest('CFOUR', 'REFERENCE', qref, **kwgs)
-
-    # qcdb/scf__d_convergence --> cfour/scf_conv
-    qopt = ropts.scroll['QCDB']['SCF__D_CONVERGENCE']
-    if qopt.disputed():
-        conv = conv_float2negexp(qopt.value)
-        #conv = conv_float2negexp(ropts.scroll['QCDB']['SCF__D_CONVERGENCE'].value)
-        ropts.suggest('CFOUR', 'SCF_CONV', conv, **kwgs)
-
-    # qcdb/scf__maxiter --> cfour/scf_maxcyc
-    ropts.suggest('CFOUR', 'SCF_MAXCYC', ropts.scroll['QCDB']['SCF__MAXITER'].value, **kwgs)
-
-    # qcdb/scf__damping_percentage --> cfour/scf_damping
-    damp = int(10 * ropts.scroll['QCDB']['SCF__DAMPING_PERCENTAGE'].value)
-    ropts.suggest('CFOUR', 'SCF_DAMPING', damp, **kwgs)
-
 
 # Philosophy break:
 #   Specification options
@@ -94,88 +42,6 @@ def muster_inherited_keywords(ropts: 'Keywords', verbose: int = 1) -> None:
 #   energy(name)            [cfour][cfour_calc_level]
 #   c4-scf                  SCF by default
 #   c4-scf                  CCSD from set {...}
-
-
-def muster_modelchem(name, dertype, ropts, verbose=1):
-    """Transform calculation method *name* and derivative level *dertype*
-    into options for cfour. While deliberately requested pieces,
-    generally |cfour__cfour_deriv_level| and |cfour__cfour_calc_level|,
-    are set to complain if contradicted ('clobber' set to True), other
-    'recommended' settings, like |cfour__cfour_cc_program|, can be
-    countermanded by keywords in input file ('clobber' set to False).
-    Occasionally, want these pieces to actually overcome keywords in
-    input file ('superclobber' set to True).
-
-    """
-    lowername = name.lower()
-    accession = 2345
-
-    if dertype == 0:
-        if lowername == 'c4-cfour':
-            pass  # permit clean operation of sandwich mode
-        else:
-            ropts.require('CFOUR', 'DERIV_LEVEL', 'ZERO', accession=accession, verbose=verbose)
-    elif dertype == 1:
-        ropts.require('CFOUR', 'DERIV_LEVEL', 'FIRST', accession=accession, verbose=verbose)
-    elif dertype == 2:
-        #ropts.require('CFOUR', 'DERIV_LEVEL', 'SECOND', accession=accession, verbose=verbose)
-        ropts.require('CFOUR', 'VIBRATION', 'EXACT', accession=accession, verbose=verbose)
-    else:
-        raise ValidationError(f"""Requested Cfour dertype '{dertype}' is not available.""")
-
-    if lowername == 'c4-cfour':
-        pass
-    elif lowername in ['c4-scf', 'c4-hf']:
-        ropts.require('CFOUR', 'CALC_LEVEL', 'SCF', accession=accession, verbose=verbose)
-
-    elif lowername == 'c4-mp2':
-        ropts.require('CFOUR', 'CALC_LEVEL', 'MP2', accession=accession, verbose=verbose)
-
-    elif lowername == 'c4-mp3':
-        ropts.require('CFOUR', 'CALC_LEVEL', 'MP3', accession=accession, verbose=verbose)
-
-    elif lowername == 'c4-mp4(sdq)':
-        ropts.require('CFOUR', 'CALC_LEVEL', 'SDQ-MP4', accession=accession, verbose=verbose)
-
-    elif lowername == 'c4-mp4':
-        ropts.require('CFOUR', 'CALC_LEVEL', 'MP4', accession=accession, verbose=verbose)
-
-    elif lowername == 'c4-cc2':
-        ropts.require('CFOUR', 'CALC_LEVEL', 'CC2', accession=accession, verbose=verbose)
-
-    elif lowername == 'c4-ccsd':
-        ropts.require('CFOUR', 'CALC_LEVEL', 'CCSD', accession=accession, verbose=verbose)
-        ropts.suggest('CFOUR', 'CC_PROGRAM', 'ECC', accession=accession, verbose=verbose)
-
-    elif lowername == 'c4-ccsd-dboc':
-        ropts.require('CFOUR', 'CALC_LEVEL', 'CCSD', accession=accession, verbose=verbose)
-        ropts.require('CFOUR', 'DERIV_LEVEL', 'FIRST', accession=accession, verbose=verbose)
-        ropts.require('CFOUR', 'DBOC', 'ON', accession=accession, verbose=verbose)
-        ropts.suggest('CFOUR', 'CC_PROGRAM', 'ECC', accession=accession, verbose=verbose)
-
-    elif lowername == 'c4-cc3':
-        ropts.require('CFOUR', 'CALC_LEVEL', 'CC3', accession=accession, verbose=verbose)
-
-    elif lowername == 'c4-ccsd(t)':
-        # Can't use (T) b/c bug in xsymcor lops it off
-        ropts.require('CFOUR', 'CALC_LEVEL', 'CCSD[T]', accession=accession, verbose=verbose)
-        ropts.suggest('CFOUR', 'CC_PROGRAM', 'ECC', accession=accession, verbose=verbose)
-
-    elif lowername == 'c4-ccsdt':
-        # TODO, CC_PROG needs defaulting on a per-reference basis
-        ropts.require('CFOUR', 'CALC_LEVEL', 'CCSDT', accession=accession, verbose=verbose)
-        ropts.suggest('CFOUR', 'CC_PROGRAM', 'ECC', accession=accession, verbose=verbose)
-
-    elif lowername == 'c4-ccsdt(q)':
-        ropts.require('CFOUR', 'CALC_LEVEL', 'CCSDT(Q)', accession=accession, verbose=verbose)
-        ropts.suggest('CFOUR', 'CC_PROGRAM', 'NCC', accession=accession, verbose=verbose)
-
-    elif lowername == 'c4-ccsdtq':
-        ropts.require('CFOUR', 'CALC_LEVEL', 'CCSDTQ', accession=accession, verbose=verbose)
-        ropts.suggest('CFOUR', 'CC_PROGRAM', 'NCC', accession=accession, verbose=verbose)
-
-    else:
-        raise ValidationError(f"""Requested Cfour computational method '{lowername}' is not available.""")
 
 
 def harvest_zmat(zmat: str) -> Molecule:
